DDCNL-8880: refactor paye connector
