+++ app/uk/gov/hmrc/tai/connectors/CompanyCarConnector.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.tai.connectors.deprecated
+package uk.gov.hmrc.tai.connectors
 
 import com.google.inject.{Inject, Singleton}
 import play.api.libs.json.Reads
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.tai.connectors.{HttpHandler, PayeUrls}
 import uk.gov.hmrc.tai.model.domain.benefits.CompanyCarBenefit
 import uk.gov.hmrc.tai.model.domain.formatters.CompanyCarBenefitFormatters
 import uk.gov.hmrc.tai.model.enums.APITypes
