+++ app/uk/gov/hmrc/tai/service/CompanyCarBenefitService.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.tai.repositories.deprecated
+package uk.gov.hmrc.tai.service
 
 import com.google.inject.{Inject, Singleton}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.tai.connectors.cache.CacheId
-import uk.gov.hmrc.tai.connectors.deprecated.CompanyCarConnector
+import uk.gov.hmrc.tai.connectors.CompanyCarConnector
 import uk.gov.hmrc.tai.model.domain.benefits.CompanyCarBenefit
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.util.MongoConstants
@@ -28,29 +27,18 @@ import uk.gov.hmrc.tai.util.MongoConstants
 import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
-class CompanyCarBenefitRepository @Inject() (
-  taiCacheRepository: TaiCacheRepository,
+class CompanyCarBenefitService @Inject() (
   companyCarConnector: CompanyCarConnector
 )(implicit ec: ExecutionContext)
     extends MongoConstants {
 
   def carBenefit(nino: Nino, taxYear: TaxYear)(implicit hc: HeaderCarrier): Future[Seq[CompanyCarBenefit]] = {
-    val cacheId = CacheId(nino)
+    val companyCarBenefits = companyCarConnector.carBenefits(nino, taxYear)
+    val version = companyCarConnector.ninoVersion(nino)
 
-    taiCacheRepository.find[Seq[CompanyCarBenefit]](cacheId, CarBenefitKey) flatMap {
-      case None =>
-        val companyCarBenefits = companyCarConnector.carBenefits(nino, taxYear)
-        val version = companyCarConnector.ninoVersion(nino)
-
-        val companyCarBenefitsWithVersion = for {
-          cc  <- companyCarBenefits
-          ver <- version
-        } yield cc.map(cc => CompanyCarBenefit(cc.employmentSeqNo, cc.grossAmount, cc.companyCars, Some(ver)))
-
-        companyCarBenefitsWithVersion.flatMap { result =>
-          taiCacheRepository.createOrUpdate(cacheId, result, CarBenefitKey).map(_ => result)
-        }
-      case Some(seq) => Future.successful(seq)
-    }
+    for {
+      cc  <- companyCarBenefits
+      ver <- version
+    } yield cc.map(cc => CompanyCarBenefit(cc.employmentSeqNo, cc.grossAmount, cc.companyCars, Some(ver)))
   }
 }
